+++ src/Composed.State.Tests/ComposeTests.cs
@@ -10,10 +10,10 @@ namespace Composed.State.Tests
     public class ComposeTests
     {
         [Fact]
-        public void UseState_NullArguments_ThrowsArgumentNullException()
+        public void UseStore_NullArguments_ThrowsArgumentNullException()
         {
-            Should.Throw<ArgumentNullException>(() => UseState<object, object>(store: null!, stateSelector: state => state, equalityComparer: EqualityComparer<object>.Default, DefaultScheduler.Instance));
-            Should.Throw<ArgumentNullException>(() => UseState(store: new TestableStore<object>(new object()), stateSelector: null!, equalityComparer: EqualityComparer<object>.Default, DefaultScheduler.Instance));
+            Should.Throw<ArgumentNullException>(() => UseStore<object, object>(store: null!, stateSelector: state => state, equalityComparer: EqualityComparer<object>.Default, DefaultScheduler.Instance));
+            Should.Throw<ArgumentNullException>(() => UseStore(store: new TestableStore<object>(new object()), stateSelector: null!, equalityComparer: EqualityComparer<object>.Default, DefaultScheduler.Instance));
         }
     }
 }